+++ resources/res_map_title.h
@@ -8,8 +8,8 @@
 extern "C" {
 #endif
 
-extern const uint8_t map_title_gfx_titlebkg_map_width;
-extern const uint8_t map_title_gfx_titlebkg_map_height;
+extern const uint16_t map_title_gfx_titlebkg_map_width;
+extern const uint16_t map_title_gfx_titlebkg_map_height;
 extern const uint16_t map_title_gfx_titlebkg_map_data[];
