+++ resources/res_map_titlef.c
@@ -1,8 +1,8 @@
 
 #include "res_map_titlef.h"
 
-const uint8_t map_titlef_gfx_titlefgn_map_width = 30;
-const uint8_t map_titlef_gfx_titlefgn_map_height = 780;
+const uint16_t map_titlef_gfx_titlefgn_map_width = 30;
+const uint16_t map_titlef_gfx_titlefgn_map_height = 780;
 const uint16_t map_titlef_gfx_titlefgn_map_data[] = {
    1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,
    1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,
