+++ resources/res_map_titlef.h
@@ -8,8 +8,8 @@
 extern "C" {
 #endif
 
-extern const uint8_t map_titlef_gfx_titlefgn_map_width;
-extern const uint8_t map_titlef_gfx_titlefgn_map_height;
+extern const uint16_t map_titlef_gfx_titlefgn_map_width;
+extern const uint16_t map_titlef_gfx_titlefgn_map_height;
 extern const uint16_t map_titlef_gfx_titlefgn_map_data[];
